pyproject.toml
[project]
name = "agent-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents[litellm]>=0.10.5",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.22",
]
